refactor: remove uncommon Q classes from top ns
